Issue [#2088323]: Treat UTC dates differently from dates w/ other timezones.
iCalcreator requires that you *not* call createTimezone($vcal, 'UTC'), because
the spec says to use zulu time rather than a UTC VTIMEZONE object.

Also fixed a copy-pasta bug that probably never manifested, since it would have
only had an effect on events that have an end date that's in a different
timezone from the start date.