Issue [#2098051]: UIDs are now persistent across multiple retrievals of the ical feed.
The previous fix that I applied for issue #2070883 was the complete wrong way to
go about fixing that problem, because it introduced a new, much more sinister
issue: the UID for an event would change every time the iCal feed got downloaded.

According to the spec, the UID is supposed to be persistent, so I re-engineered
the algoithm that creates UIDs for the Fields plugin so it will create the same
UIDs as the Entity plugin does.