Issue #2107931: The Fields plugin now supports "First populated Date field".
Rather than defaulting an an illegal value, the Fields plugin's settings
form now offers "First populated Date field" as the default option for
"Date field". Using this option will tell Date iCal to look through all the
fields in the View until it finds a populated Date field, and use that field.

So, if there are two Date fields specified in the view's FIELDS setting, and
at runtime the value of the first one is empty (e.g. because the node
doesn't use that field), then the second Date field's value will be used.