Issue #2127717: Parsed single-day All Day events are now given an end date.
As jakemonO pointed out, the Calendar will not display All Day events which have no
end date value, which means that Date iCal's method of leaving off the end date for
single-day All Day events was a really bad idea. Fortuantely, setting the end date
to the same day as the start date makes the date node display correctly for single-
day events, which was the reason that the end date was being left off in the first
place. This means that every All Day event can simply have it's end date rewound by
one day, and all is well in both the node display and the Calendar.