Issue #2133483: The rename for DateiCalFeedsParser.inc didn't get pushed to git.
Because OSX uses a case-insensitive file system, git didn't pick up the change
I made when I re-wrote the parser. In retrospect, I probably should have chosen
a totally new name for the feeds parser, but the existing filename *was* the
most appropriate name for the new class, just with different capitalization.