Initial Date iCal 3.1-dev commit. Fixed multi-property RDATEs and EXDATEs.
The Date Repeat module’s date_repeat_build_dates() function does not support
RDATEs and EXDATEs defined as separate properties, as well as being a bit
buggy in other ways. So I copy-pasta’d it and made the necessary fixes.

I wanted to add support for EXRULEs, but it’s more complicated than I
expected. So since no one has actually asked for it, I left it as a TODO.