Issue #2207173: RRULEs with non-UTC UNTIL values are now supported.
According to the iCal spec, the UNTIL value of an RRULE must be specified in
UTC. But certain iCal feed creation tools don't respect this rule, and
instead specify the UNTIL value in local time. Since all the other code that
Date iCal utilizes to parse RRULEs expects the UNTIL value to be in UTC,
this go badly when it's not.

In order to get around this issue, I've added an option to the iCal Parser
settings, which allows users to tell Date iCal that they expect the UNTIL
values in their feeds' RRULEs to not be in UTC. With that option enabled,
Date iCal will treat the UNTIL date as being in the same timezone as the
event's DTSTART.