Issue #2466565: Added a workaround to make VFREEBUSY componenets importable.
Because the iCal spec states that VFREEBUSY components can't have the
SUMMARY property, Date iCal will now fall back to the COMMENT property
for the node title. Unfortuantely, this is the best we can do, becaise
iCalcreator completely leaves out the SUMMARY property on VFREEBUSY
componeents, so there's no way to recover it.