Issue #2625002: Fixed warnings thrown during RRULE parsing under some setups.
When I wrote up a fix for one of the changes made in the newest version of Feeds,
I put the fix a few lines after it should have gone. This resulted in blank
RRULEs going undetected when those versions of Feeds are installed, making the
parsing code attempt to parse a blank string, which threw warnings and notices.